Rework the code + cleanup + optimize
